Task: Логика проверки ответа Логика перехода от одного вопроса к следующему
