+++ app/src/main/java/com/bignerdranch/chemcraft/App.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import com.bignerdranch.chemcraft.cards.di.cardsModule
 import com.bignerdranch.chemcraft.items_in_lesson.di.itemsModule
 import com.bignerdranch.chemcraft.lessons.di.lessonsModule
-import com.bignerdranch.chemcraft.task_for_card.di.taskModel
+import com.bignerdranch.chemcraft.lesson_task.di.taskModel
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
