+++ app/src/main/java/com/bignerdranch/chemcraft/MainMenuActivity.kt
@@ -5,22 +5,18 @@ import android.util.Log
 
 
 import androidx.appcompat.app.AppCompatActivity
-import com.bignerdranch.chemcraft.network.old.FirebaseAdminManager
+import com.bignerdranch.chemcraft.network.admin_manager.FirebaseAdminManager
 import com.bignerdranch.chemcraft.databinding.ActivityMainBinding
-import com.bignerdranch.chemcraft.network.old.CardToServerModel
 import com.bignerdranch.chemcraft.lessons.ui.LessonActivity
-import com.bignerdranch.chemcraft.network.old.TestToServerModel
-import com.bignerdranch.chemcraft.task_for_card.ui.TaskActivity
-import com.bignerdranch.chemcraft.task_for_card.domain.model.TaskModel
-import com.google.firebase.Firebase
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.firestore
+import com.bignerdranch.chemcraft.network.admin_manager.MokeCardToFirebase
+import com.bignerdranch.chemcraft.network.admin_manager.models.TaskToServerModel
+import com.bignerdranch.chemcraft.lesson_task.ui.TaskActivity
+import com.bignerdranch.chemcraft.local_storage.FavoriteActivity
+import com.bignerdranch.chemcraft.network.admin_manager.MokeTaskToFirebase
 
 
 class MainMenuActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
-    private lateinit var db: FirebaseFirestore
-
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -41,146 +37,15 @@ class MainMenuActivity : AppCompatActivity() {
             startActivity(Intent(this, TaskActivity::class.java))
         }
 
-        val cardItemsMoke = listOf(
-            mapOf(
-                "type" to "text",
-                "content" to "Более трёх с половиной тысяч лет назад на юге Балканского полуострова, в Европе, жили греческие племена. Значительную часть Греции — страны, которую они населяли, — покрывают крутые и обрывистые горы. Эти горы делят территорию на изолированные друг от друга области.\n" +
-                    "\n" +
-                    "Горный рельеф имел как плюсы, так и минусы. Горы надёжно защищали Элладу (Древнюю Грецию) от внешних захватчиков. На протяжении истории страна десятилетиями спокойно развивалась без угрозы вражеского вторжения, что в условиях Древнего мира было редкостью. Однако горы также затрудняли общение между различными областями. От одной деревни до другой иногда было всего несколько километров по прямой, но скалы и ущелья растягивали путь на целый день, а то и больше. В таких условиях объединение в единое государство было крайне затруднено.\n" +
-                    "\n" +
-                    "Крупных и полноводных рек в Греции нет. Со всех сторон, кроме северной, страну омывают моря. Платон говорил, что греки живут по берегам Средиземного моря, как лягушки по берегам болота. В море много островов, и самый большой из них — Крит. Именно на Крите возникла древнейшая цивилизация на территории Европы — Минойская.\n"
-            ),
-            mapOf("type" to "img", "content" to "https://res.cloudinary.com/dmgaamkur/image/upload/v1752615365/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA1_itvuun.png"),
-
-            mapOf("type" to "text",
-                "content" to "С древних времён (III—II тысячелетие до н. э.) на Крите проживал народ негреческого происхождения. Жителей этого острова сегодня называют критянами или минойцами — в честь легендарного царя Миноса.\n" +
-                    "\n" +
-                    "Остров Крит занимал выгодное положение на морском пути между Грецией и Египтом. В конце III тысячелетия до н. э. на острове начали формироваться небольшие города-государства, в центре каждого из которых располагался дворец правителя. Большую часть населения составляли землевладельцы, которые обрабатывали свои участки земли.\n" +
-                    "\n" +
-                    "Критская цивилизация достигла расцвета в XVI — первой половине XV века до н. э. В этот период произошло объединение городов-государств под властью царей города Кносса. Огромный Кносский дворец насчитывал сотни помещений и проходов, а также включал залы для проведения религиозных обрядов.\n"),
-
-            mapOf("type" to "img", "content" to "https://res.cloudinary.com/dmgaamkur/image/upload/v1752617390/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA2_bg3w5x.png"),
-
-            mapOf("type" to "img", "content" to "https://res.cloudinary.com/dmgaamkur/image/upload/v1752617424/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA3_x4pnv8.png"),
-
-            mapOf("type" to "text",
-                "content" to "Дворцовые стены были украшены фресками на различные темы. При раскопках во дворце также были найдены многочисленные глиняные таблички с крито-минойским письмом, которое до сих пор не расшифровано.\n" +
-                    "\n" +
-                    "Самым известным кносским царём был Минос. По легенде, искусный изобретатель Дедал создал при Миносе огромный и запутанный Лабиринт, в котором обитало мифическое чудовище — Минотавр.\n" +
-                    "\n" +
-                    "Критские правители обладали мощным военным флотом и господствовали в Эгейском море, торгуя с Египтом, финикийскими городами и другими странами Средиземноморья.\n" +
-                    "\n" +
-                    "Однако на пике своего развития критская цивилизация была уничтожена природной катастрофой — катастрофическим извержением вулкана, которое произошло в середине XV века до н. э. Города и дворцы на острове были разрушены, а население бежало. Спустя время на Крит переселились ахейские греки. Они заимствовали у критян некоторые религиозные обряды, живопись и письменность. Поэтому ранний период истории Древней Греции учёные называют Крито-Микенской цивилизацией.\n"
-            ),
-
-            mapOf("type" to "img", "content" to "https://res.cloudinary.com/dmgaamkur/image/upload/v1752617505/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA5_gipatd.png"),
-
-            mapOf("type" to "img", "content" to "https://res.cloudinary.com/dmgaamkur/image/upload/v1752617503/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA4_drlafj.jpg"),
-
-            mapOf("type" to "text",
-                "content" to "Легенда о Тесее и Минотавре\n" +
-                        "Однажды царь Кносса Минос отправил своего сына Андрогея в Афины, чтобы тот принял участие в спортивных состязаниях. Андрогей одержал победу над афинянами, но царь Афин Эгей, охваченный завистью, приказал его убить.\n" +
-                        "\n" +
-                        "Разгневанный Минос выступил с флотом против Афин, захватил город и вынудил его жителей платить тяжёлую дань. Каждые девять лет афиняне должны были отправлять на Крит семь юношей и семь девушек, которых приносили в жертву Минотавру — чудовищному существу с телом человека и головой быка. Это существо обитало в Лабиринте — запутанном сооружении, полном извилистых ходов, выбраться из которого было невозможно.\n" +
-                        "\n" +
-                        "Когда пришло время новой жертвы, Эгей предложил жребий, чтобы определить, чьи дети отправятся на смерть. Однако народ возмутился: почему сам царь не участвует в жеребьёвке? Услышав эти упрёки, сын Эгея, Тесей, добровольно вызвался отправиться на Крит, чтобы сразиться с Минотавром. Перед отплытием отец просил его, если он вернётся победителем, заменить чёрный парус корабля на белый, чтобы издалека было видно радостную весть.\n" +
-                        "\n" +
-                        "Когда афиняне прибыли на Крит, их привели во дворец Миноса. Там царская дочь Ариадна, влюбившись в Тесея, решила помочь ему. По совету архитектора Дедала, создавшего Лабиринт, она тайком передала герою меч и клубок нитей. Ариадна объяснила, что нужно привязать конец нити у входа в Лабиринт и разматывать клубок, двигаясь по коридорам, чтобы найти путь обратно.\n" +
-                        "\n" +
-                        "Вместе с афинскими юношами и девушками Тесей вошёл в запутанные ходы Лабиринта. Внезапно раздался яростный рёв — навстречу выбежал Минотавр. Чудовище бросилось на героя, готовясь пронзить его рогами, но Тесей ловко уклонился и смертельно ранил его мечом. Затем, следуя нити Ариадны, афиняне выбрались наружу и поспешно покинули Крит. Однако по пути домой Тесей, увлечённый радостью победы, забыл сменить чёрный парус на белый. Увидев траурный знак, Эгей решил, что сын погиб, и, охваченный горем, бросился в море со скалы. С тех пор это море стало называться Эгейским.\n"
-            ),
-
-            mapOf(
-                "type" to "img",
-                "content" to " https://res.cloudinary.com/dmgaamkur/image/upload/v1752617625/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA6_zec3bn.png"
-            ),
-
-            mapOf(
-                "type" to "text",
-                "content" to "Легенда о Дедале и Икаре\n" +
-                        "Дедал был выдающимся афинским мастером, архитектором и изобретателем. Однажды он оказался на службе у критского царя Миноса, для которого построил знаменитый Лабиринт — запутанное сооружение, созданное для заключения Минотавра.\n" +
-                        "\n" +
-                        "Однако позднее царь разгневался на Дедала, заподозрив, что именно он помог Тесею одолеть чудовище и сбежать. Чтобы мастер не мог покинуть остров и передать свои знания врагам, Минос заточил его вместе с сыном Икаром в высокую башню на Крите.\n" +
-                        "\n" +
-                        "Но Дедал не был бы великим изобретателем, если бы не нашёл способ выбраться. Видя, что сушу и море охраняют, он решил бежать по воздуху. Собрав птичьи перья и скрепив их воском, мастер создал две пары крыльев — для себя и для сына. Перед полётом он строго наказал Икару: не подниматься слишком высоко, чтобы солнце не растопило воск, и не лететь слишком низко, чтобы влага не намочила перья.\n" +
-                        "\n" +
-                        "Они взлетели, оставив Крит позади. Дедал уверенно парил в воздухе, но юный Икар, охваченный восторгом, забыл наставления отца. Восхищённый своей свободой, он поднялся всё выше и выше. Солнечные лучи раскалили воск, и крылья Икара начали распадаться. В отчаянии он пытался удержаться в воздухе, но вскоре рухнул в море и утонул.\n"
-            ),
-
-            mapOf(
-                "type" to "img",
-                "content" to " https://res.cloudinary.com/dmgaamkur/image/upload/v1752618105/%D0%A0%D0%B8%D1%81%D1%83%D0%BD%D0%BE%D0%BA1_suyig7.png"
-            )
-
-        )
 
 
+
+///////////// создаем урок
         // Данные
         val lessonData: HashMap<String, Any> = hashMapOf(
             "description" to "333",
             "name" to "Древний Рим",
         )
-
-        val testData  = listOf(
-            TaskModel(
-                questionText = "Установите правильную последовательность событий, предшествующих образованию зиготы у человека. Запишите в таблицу соответствующую последовательность цифр.\n" +
-                        "1) выход яйцеклетки из фолликула в маточную трубу\n" +
-                        "2) накопление питательных веществ в яйцеклетке\n" +
-                        "3) проникновение ядра сперматозоида в яйцеклетку\n" +
-                        "4) формирование первичного фолликула\n" +
-                        "5) окончательное прохождение первого деления мейоза яйцеклеткой",
-                imgUrl = "",
-                correctAnswer = "42513",
-                maxScore = 1,
-                type = 1
-
-            ),
-            TaskModel(
-                questionText = "16531. Расположите в хронологической последовательности исторические события. Запишите цифры, которыми обозначены исторические события, в правильной последовательности в поле ответа.\n" +
-                        "1) провозглашение Германской империи\n" +
-                        "2) Чигиринские походы\n" +
-                        "3) создание Верховного тайного совета",
-                imgUrl = "",
-                correctAnswer = "231",
-                maxScore = 1,
-                type = 1
-
-            ),
-            TaskModel(
-                questionText = "16530. Установите соответствие между событиями и годами: к каждой позиции первого столбца подберите соответствующую позицию из второго столбца.\n" +
-                        "СОБЫТИЯ\n" +
-                        "\n" +
-                        "А) заключение письменного договора Руси с Византией\n" +
-                        "Б) дарование конституции Царству Польскому\n" +
-                        "В) Медный бунт\n" +
-                        "Г) возведение Берлинской стены\n" +
-                        "ГОДЫ\n" +
-                        "\n" +
-                        "1) 911 г.\n" +
-                        "2) 1097 г.\n" +
-                        "3) 1662 г.\n" +
-                        "4) 1815 г.\n" +
-                        "5) 1945 г.\n" +
-                        "6) 1961 г.",
-                imgUrl = "",
-                correctAnswer = "1436",
-                maxScore = 1,
-                type = 1
-
-            )
-        )
-
-        val testToServerData = TestToServerModel (
-            testContent = testData
-        )
-
-        val cardsToServerData = CardToServerModel(
-            imageUrl = "Тут должна быть картинка",
-            important = false,
-            title = "Заголовок мок",
-            cardItems = cardItemsMoke // биндим однотимным текстом и картинками (мок)
-        )
-
 
         // создаем урок
         binding.adminButtonOne.setOnClickListener {
@@ -194,18 +59,27 @@ class MainMenuActivity : AppCompatActivity() {
                 )
 
         }
+///////////
 
 
-            // создать карточку в уроке
+///////////////  Эта логика по идее должна быть в классе data
+
+        val lessonId_CardToLesson: String = "tPG7SRYJRV64l9dcmjDa"
+
+
         binding.adminButtonTwo.setOnClickListener {
 
-            val lessonId: String = "jcOHbkq3m6YiQnQKyXVW"
+
 
             FirebaseAdminManager.addCardToLesson(
-                lessonId = lessonId,
-                cardModel = cardsToServerData,
+                lessonId = lessonId_CardToLesson,
+                cardModel = MokeCardToFirebase.cardsToServerData,
                 onSuccess = { result ->
-                    Log.d("Карточка  добавлена в урок", "Карточка: ${result.id} в $lessonId")
+                    Log.d("Карточка  добавлена в урок", "Карточка: ${result.id} в $lessonId_CardToLesson")
+                    MokeCardToFirebase.cardItemsMoke.forEach{
+                        Log.d("Данные загруженной карточки", "$it")
+                    }
+
                 },
                 onFailure = { }
             )
@@ -213,30 +87,38 @@ class MainMenuActivity : AppCompatActivity() {
 
         }
 
+///////////////////
+        /////////////////////
+
+
+
+////////////////////
+
+        val lessonIdtestToServerData: String = "tPG7SRYJRV64l9dcmjDa"
+        val cardIdtestToServerData: String = "TQabQdCAeEMnOg0c0DIY"
+
+
+        val testToServerData = TaskToServerModel (
+            testContent = MokeTaskToFirebase.testData
+        )
+
+        // добавить тест в карточку
         binding.adminButtonThree.setOnClickListener {
 
-            val lessonId: String = "lO2juSRIkYm7Wzb0og9G"
-            val cardId: String = "Hoa6J343cc9wZ1Bx7DRZ"
-
-            //
-                    FirebaseAdminManager.addTestToCard(
-                        lessonId = lessonId,
-                        cardId = cardId,
-                        testData = testToServerData,
-                        onSuccess = { result ->
-                            Log.d("Тест загружен", "Путь: тест: ${result.id} -> \n карточка: $cardId \n -> урок: $lessonId ")
-                        }
+            FirebaseAdminManager.addTestToCard(
+                lessonId = lessonIdtestToServerData,
+                cardId = cardIdtestToServerData,
+                testData = testToServerData,
+                onSuccess = { result ->
+                    Log.d(
+                        "Тест загружен",
+                        "ID теста: ${result.id} <-  \n <- карточка: $lessonIdtestToServerData \n <- урок: $cardIdtestToServerData "
                     )
-
+                }
+            )
         }
 
 
 
-
-        db = Firebase.firestore
-
-
     }
-
-
 }
