+++ app/src/main/java/com/bignerdranch/chemcraft/lesson_task/data/model/ClientGetTasks.kt
@@ -1,0 +1,7 @@
+package com.bignerdranch.chemcraft.lesson_task.data
+
+import com.bignerdranch.chemcraft.network.Response
+
+interface ClientGetTasks {
+    suspend fun getTasksById(lessonId: String, cardId: String): Response
+}
