+++ app/src/main/java/com/bignerdranch/chemcraft/lesson_task/data/model/TasksResponse.kt
@@ -1,0 +1,8 @@
+package com.bignerdranch.chemcraft.lesson_task.data.model
+
+import com.bignerdranch.chemcraft.network.Response
+
+class TasksResponse(
+    val result: List<Map<String, Any>>
+): Response() {
+}
