+++ app/src/main/java/com/bignerdranch/chemcraft/lesson_task/ui/TaskState.kt
@@ -1,0 +1,17 @@
+package com.bignerdranch.chemcraft.lesson_task.ui
+
+import com.bignerdranch.chemcraft.lesson_task.domain.model.TaskModel
+
+interface TaskState {
+
+    data object Loading : TaskState
+
+    data class Error(
+        val errorMessage: String
+    ) : TaskState
+
+    data class Content(
+        val items: List<TaskModel>
+    ) : TaskState
+
+}
