+++ app/src/main/java/com/bignerdranch/chemcraft/lesson_task/ui/answer_screen/AllTestResultModel.kt
@@ -1,0 +1,8 @@
+package com.bignerdranch.chemcraft.lesson_task.model
+
+data class AllTestResultModel(
+    val correct: String,
+    val userAnswer: String,
+    val getScore: Int
+) {
+}
