+++ app/src/main/java/com/bignerdranch/chemcraft/lesson_task/ui/answer_screen/AllTestResultViewHolder.kt
@@ -1,0 +1,26 @@
+package com.bignerdranch.chemcraft.lesson_task.ui.answer_screen
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+import com.bignerdranch.chemcraft.lesson_task.model.AllTestResultModel
+
+class AllTestResultViewHolder(itemView: View): RecyclerView.ViewHolder(itemView) {
+
+//    private val cardResultNumber: TextView = itemView.findViewById(R.id.testTaskNumber)
+//    private val checkAnswer: TextView = itemView.findViewById(R.id.checkAnswer)
+//    private val userAnswer: TextView = itemView.findViewById(R.id.userAnswer)
+
+
+    fun bind(model: AllTestResultModel, position: Int) {
+
+
+//        cardResultNumber.text = (position + 1).toString()
+//        checkAnswer.text = model.correct
+//        userAnswer.text = model.userAnswer
+
+    }
+
+
+
+
+}
