+++ app/src/main/java/com/bignerdranch/chemcraft/lessons/ui/LessonActivity.kt
@@ -53,6 +53,8 @@ class LessonActivity : AppCompatActivity() {
         viewModel.getLessons()
 
 
+
+
     }
 
     private fun render(state: LessonsState) {
