+++ app/src/main/java/com/bignerdranch/chemcraft/network/admin_manager/MokeCardToFirebase.kt
@@ -1,0 +1,101 @@
+package com.bignerdranch.chemcraft.network.admin_manager
+
+import com.bignerdranch.chemcraft.network.admin_manager.models.CardToServerModel
+
+object MokeCardToFirebase {
+
+    var textContent = "Упадок греческих полисов\n" +
+            "Недолго длился расцвет греческих городов. Ещё при жизни Перикла в самой Греции начались междоусобные войны. Сначала — между Афинами и Спартой, которые соперничали за господство над всей Элладой. Эта война, вошедшая в историю как Пелопоннесская, закончилась победой Спарты.\n" +
+            "Однако вскоре против Спарты выступили другие греческие полисы. Казалось, войнам не будет конца... Настали страшные времена. Лилась кровь — греки убивали друг друга. Воины разрушали селения, вырубали оливковые рощи и виноградники, вытаптывали поля с пшеницей и ячменём. Разорялись земледельцы и ремесленники, а их имущество скупали богачи.\n" +
+            "Постоянные войны между городами и внутренние конфликты между бедняками и богатыми ослабляли Элладу. Тем временем усиливалось Македонское царство — северный сосед Греции.\n"
+
+
+    var textContent2 = "§43. Города Эллады подчиняются Македонии\n" +
+            "1. Македонский царь Филипп стремится покорить соседей\n" +
+            "В середине IV (4-го) века до н. э. во главе Македонии — маленькой горной страны на севере Балканского полуострова — встал умный и энергичный царь Филипп II. Македонские цари находились под сильным влиянием греческой культуры. Они строили театры, приглашали эллинских поэтов, художников, философов. По просьбе царя Филиппа учителем его сына Александра стал крупнейший философ Греции — Аристотель. Многие знатные македоняне свободно говорили и писали по-гречески.\n" +
+            "\n" +
+            "Филипп создал большую и боеспособную армию. Македонская конница состояла из знати, а пехота набиралась из пастухов и земледельцев. Каждый пехотинец получал шлем, кожаный панцирь, круглый щит, короткий меч и длинное копьё — сариссу, длиной до 6 метров. Фаланга строилась в 16 рядов. В бою первые 6 рядов держали копья обеими руками, направляя их на врага. Передний ряд выставлял копья вперёд, а воины сзади клали их на плечи товарищей впереди. Если враг заходил фаланге в тыл, задние ряды разворачивались, и строй становился неприступным со всех сторон. Недаром говорили, что фаланга похожа на ощетинившегося железом зверя, к которому опасно приближаться.\n" +
+            "\n" +
+            "Армия Филиппа использовала метательные орудия, с помощью которых в противника метали камни, брёвна и огромные стрелы. Эти орудия устанавливали в многоэтажных деревянных башнях на колёсах. Сотни воинов катили такие башни к стенам осаждённых городов. Они грохотали и наводили ужас на защитников.\n" +
+            "Покоряя соседние земли, Филипп рассчитывал не только на силу оружия. Он умело сталкивал между собой своих противников, нередко добивался успеха подкупом. Царь с издёвкой говорил: «Даже в самой неприступной крепости найдётся калитка, в которую пройдёт осёл, нагруженный золотом».\n"
+
+    var textContent3 = "2. Потеря Грецией независимости\n" +
+            "Филипп начал подчинять себе греческие города один за другим. Эти полисы были ослаблены ожесточённой, часто кровавой борьбой между богатыми и бедными. Афинский писатель Исократ утверждал: «Врага боятся меньше, чем собственных сограждан. Богатые скорее бросят своё имущество в море, чем отдадут его бедным, а бедняки больше всего мечтают ограбить богатых». Исократ видел спасение родины в добровольном подчинении Филиппу. Он призывал его объединить всех эллинов и вместе двинуться походом против пер\u00ADсидского царя. \n" +
+            "\n" +
+            "Но были и такие греки, которые выше всего ставили свободу и независимость своего родного города. Ярым и непримиримым противником Филиппа в Афинах стал знаменитый оратор Демосфен. Разъезжая по городам Эллады и произнося гневные речи, он сплотил ряд полисов для борьбы с македонским царём.\n" +
+            "Решающее сражение произошло в 338 году до н. э. близ города Херонеи. Демосфен сражался в боевом строю как простой воин. Хотя македоняне превосходили греков численностью, битва была упорной. Под угрозой оказалась сама жизнь Филиппа. Тогда царь прибег к военной хитрости: притворно отступил, увлекая за собой афинские отряды. Греки не заподозрили подвоха и бросились в погоню. В этот момент восемнадцатилетний Александр во главе конницы нанёс мощный удар по флангу греческой армии. Видя успехи сына, Филипп тут же повернул свои отряды и добил растерянного противника. Поле битвы покрылось телами павших. Вне себя от радости, Филипп устроил пир прямо среди погибших. \n" +
+            "\n" +
+            "Однако Филипп не стал объявлять себя царём Греции. Вместо этого греческие полисы принесли ему клятву верности и признали его верховным военным предводителем. Так был образован Коринфский союз между Македонией и греческими городами. Немедленно началась подготовка македонян и греков на персов.\n"
+
+
+    var textContent4 = "3. Приход к власти Александра\n" +
+            "Все приготовления были завершены, но осуществить поход Филиппу не удалось. В Македонии шли празднества по слу­чаю свадьбы царской дочери. Во время праздника Филипп направлялся в театр в сопровождении друзей и телохранителей. Внезапно на него напал один из приближённых. Из-под одежды он выхватил изогнутый клинок и смертельно ранил царя. Убийца попытался бежать, но стража настигла и заколола его. Причины убийства так и остались загадкой.\n" +
+    "Новоe македонское войско тут же провозгласило царём 20-летнего Александра. Александр Македонский решительно продолжил дело отца: он возглавил объединённые силы греков и македонян и начал великий поход против Персидской державы — самого могущественного государства своего времени.\n"
+
+
+
+    var textContent5 = "Древний писатель Плутарх рассказывает о Демосфене\u000BДемосфен был сыном владельца оружейной мастерской. С семи лет он осиротел, унаследовав от отца большое состояние, но опекуны расхитили все его деньги. " +
+            "С детских лет Демосфен мечтал стать оратором. Это было непросто: он рос болезненным ребёнком, имел слабый " +
+            "голос. Забросив мальчишеские забавы, Демосфен целиком отдался упражнениям в красноречии. Он уединялся в дальней " +
+            "комнате на два-три месяца, выбрив себе половину головы, чтобы от стыда невозможно было выйти наружу. Голос укреплял " +
+            "тем, что, не переводя дыхания, произносил длинные фразы. Неясный выговор он одолевал, вкладывая в рот камешки и так " +
+            "читая на память отрывки из поэтических произведений. Демосфен выступил в суде против своих опекунов и добился их осуждения.\n"
+
+
+    var urlContent = "https://ik.imagekit.io/evhsrjlhs/image2.png?updatedAt=1754558204018"
+    var urlContent2 = "https://ik.imagekit.io/evhsrjlhs/image3.jpeg?updatedAt=1754558196526"
+    var urlContent3 = "https://ik.imagekit.io/evhsrjlhs/image2.png?updatedAt=1754558204018"
+    var urlContent4 = "https://ik.imagekit.io/evhsrjlhs/image8.jpeg?updatedAt=1754558196487"
+    var urlContent5 = "https://ik.imagekit.io/evhsrjlhs/image9.png?updatedAt=1754558206450"
+
+
+
+    var urlContent31 = "https://ik.imagekit.io/evhsrjlhs/image5.jpeg?updatedAt=1754558201807"
+    var urlContent32 = "https://ik.imagekit.io/evhsrjlhs/image6.png?updatedAt=1754558203043"
+    var urlContent33 = "https://ik.imagekit.io/evhsrjlhs/image4.jpeg?updatedAt=1754558196517"
+
+    var urlContent41 = "https://ik.imagekit.io/evhsrjlhs/image7.jpeg?updatedAt=1754558196799"
+
+    var urlContent51 = "https://ik.imagekit.io/evhsrjlhs/image10.png?updatedAt=1754558196747"
+
+
+
+
+
+
+
+    // основное тело запроса ->
+
+    var cardItemsMoke = listOf(
+
+        mapOf( "type" to "text",  "content" to textContent),
+        mapOf( "type" to "img", "content" to urlContent),
+
+        mapOf( "type" to "text",  "content" to textContent2),
+        mapOf( "type" to "img", "content" to urlContent2),
+
+        mapOf( "type" to "text",  "content" to textContent3),
+        mapOf( "type" to "img", "content" to urlContent31),
+        mapOf( "type" to "img", "content" to urlContent32),
+        mapOf( "type" to "img", "content" to urlContent33),
+
+        mapOf( "type" to "text",  "content" to textContent4),
+        mapOf( "type" to "img", "content" to urlContent4),
+        mapOf( "type" to "img", "content" to urlContent41),
+
+        mapOf( "type" to "text",  "content" to textContent5),
+        mapOf( "type" to "img", "content" to urlContent5),
+        mapOf( "type" to "img", "content" to urlContent51),
+    )
+
+
+    val cardsToServerData = CardToServerModel(
+        imageUrl = "Тут должна быть картинка",
+        important = false,
+        title = "Города Эллады подчиняются Флпу II",
+        cardItems = cardItemsMoke // биндим однотимным текстом и картинками (мок)
+    )
+
+
+}
+
