+++ app/src/main/java/com/bignerdranch/chemcraft/network/admin_manager/models/CardToServerModel.kt
@@ -1,0 +1,8 @@
+package com.bignerdranch.chemcraft.network.admin_manager.models
+
+data class CardToServerModel(
+    val imageUrl: String = "",
+    val important: Boolean = false,
+    val title: String = "",
+    val cardItems: List<Map<String, String>>
+)
