+++ app/src/main/java/com/bignerdranch/chemcraft/network/admin_manager/models/TaskToServerModel.kt
@@ -1,0 +1,10 @@
+package com.bignerdranch.chemcraft.network.admin_manager.models
+
+
+import com.bignerdranch.chemcraft.lesson_task.domain.model.TaskModel
+
+data class TaskToServerModel(
+    var testContent: List<TaskModel> = emptyList(),
+    var name: String = ""
+) {
+}
