--- app/src/main/java/com/bignerdranch/chemcraft/network/old/CardToServerModel.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.chemcraft.network.old
-
-data class CardToServerModel(
-    val imageUrl: String = "",
-    val important: Boolean = false,
-    val title: String = "",
-    val cardItems: List<Map<String, String>>
-)
