--- app/src/main/java/com/bignerdranch/chemcraft/network/old/TestToServerModel.kt
@@ -1,10 +1,0 @@
-package com.bignerdranch.chemcraft.network.old
-
-
-import com.bignerdranch.chemcraft.task_for_card.domain.model.TaskModel
-
-data class TestToServerModel(
-    var testContent: List<TaskModel> = emptyList(),
-    var name: String = ""
-) {
-}
