--- app/src/main/java/com/bignerdranch/chemcraft/task_for_card/data/model/TasksResponse.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.chemcraft.task_for_card.data.model
-
-import com.bignerdranch.chemcraft.network.Response
-
-class TasksResponse(
-    val result: List<Map<String, Any>>
-): Response() {
-}
