--- app/src/main/java/com/bignerdranch/chemcraft/task_for_card/domain/api/GetTasksInteractor.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.chemcraft.task_for_card.domain.api
-
-import com.bignerdranch.chemcraft.Resource
-import com.bignerdranch.chemcraft.task_for_card.domain.model.TaskModel
-
-interface GetTasksInteractor {
-    suspend fun getTasks(lessonId: String, cardsId: String): Resource<List<TaskModel>>
-}
