--- app/src/main/java/com/bignerdranch/chemcraft/task_for_card/domain/model/TaskModel.kt
@@ -1,10 +1,0 @@
-package com.bignerdranch.chemcraft.task_for_card.domain.model
-
-data class TaskModel(
-    var questionText: String = "",
-    var imgUrl: String = "",
-    var correctAnswer: String = "",
-    var maxScore: Int = 0,
-    var type: Int = 1 // лучше заменить на seald еализацию
-) {
-}
