--- app/src/main/java/com/bignerdranch/chemcraft/task_for_card/ui/answer_screen/AllTestResultModel.kt
@@ -1,8 +1,0 @@
-package com.bignerdranch.chemcraft.task_for_card.model
-
-data class AllTestResultModel(
-    val correct: String,
-    val userAnswer: String,
-    val getScore: Int
-) {
-}
