+++ app/src/main/res/layout/activity_all_test_result.xml
@@ -6,16 +6,18 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".task_for_card.AllTestResult">
+    tools:context=".lesson_task.AllTestResult">
 
 
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/testResultRecycleView"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         android:clipToPadding="false"
         android:padding="8dp"
         android:overScrollMode="never"
