+++ app/src/main/res/layout/activity_favorite.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavoriteActivity">
+    tools:context=".local_storage.FavoriteActivity">
 
     <TextView
         android:id="@+id/my_list_is_empty"
