+++ app/src/main/res/layout/item_all_test_result.xml
@@ -1,15 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
+    android:layout_width="200dp"
+    android:layout_height="200dp"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
 
 
     <com.google.android.material.card.MaterialCardView
         android:layout_margin="5dp"
-        android:layout_width="200dp"
-        android:layout_height="200dp"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        app:layout_constraintDimensionRatio="1:1"
         app:cardElevation="2dp"
         app:cardCornerRadius="20dp"
         app:cardUseCompatPadding="true"
@@ -26,73 +27,6 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent">
-
-            <TextView
-                android:id="@+id/testTaskNumber"
-                android:layout_width="0dp"
-                android:layout_height="0dp"
-                android:gravity="center"
-                android:text="Задание $$"
-                android:textSize="20sp"
-                app:layout_constraintBottom_toTopOf="@+id/guideLineInsideCardOne"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintHorizontal_bias="0.0"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintVertical_bias="0.0" />
-
-            <androidx.constraintlayout.widget.Guideline
-                android:id="@+id/guideLineInsideCardOne"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="horizontal"
-                app:layout_constraintGuide_percent="0.25" />
-
-
-            <TextView
-                android:id="@+id/checkAnswer"
-                android:layout_width="0dp"
-                android:layout_height="0dp"
-                android:textSize="10sp"
-                android:text="Верный ответ:"
-                app:layout_constraintTop_toBottomOf="@id/guideLineInsideCardOne"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintBottom_toTopOf="@id/guideLineInsideCardTwo"
-                android:layout_margin="10dp"/>
-
-            <androidx.constraintlayout.widget.Guideline
-                android:id="@+id/guideLineInsideCardTwo"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="horizontal"
-                app:layout_constraintGuide_percent="0.6" />
-
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="1dp"
-                app:layout_constraintBottom_toTopOf="@id/guideLineInsideCardTwo"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                android:backgroundTint="@color/dark_gray"
-                android:layout_marginStart="30dp"
-                android:layout_marginEnd="30dp"
-
-                />
-
-            <TextView
-                android:id="@+id/userAnswer"
-                android:layout_width="0dp"
-                android:layout_height="0dp"
-                android:textSize="10sp"
-                android:text="Твой ответ:"
-                app:layout_constraintTop_toBottomOf="@id/guideLineInsideCardTwo"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                android:layout_margin="10dp"/>
-
-
 
 
         </androidx.constraintlayout.widget.ConstraintLayout>
